feat: createStudent, updateStudent

+++ src/controller/StudentController.ts
@@ -9,7 +9,8 @@ export class StudentController{
         this.studentService = studentService;
     }
 
-    private async getAllStudents (req:Request, res:Response){
+
+    async getAllStudents (req:Request, res:Response){
         try{
             const students = await this.studentService.getAllStudents();
             res.status(200).json(students);
@@ -20,10 +21,11 @@ export class StudentController{
         }
     };
 
-    private async getStudentsById (req:Request, res:Response){
+     async getStudentsById (req:Request, res:Response){
         const id = Number(req.params.id);
         if (isNaN(id)){
             res.status(400).json({message:'id invalide'});
+            return;
         }
         try {
             const students = await this.studentService.getStudentsById(id)
@@ -34,10 +36,46 @@ export class StudentController{
             res.status(404).json({message: 'id pas trouvé'})
         }
     };
-    private async deleteStudent (req:Request, res:Response){
+
+     async createStudent (req:Request, res:Response){
+        const {firstname, lastname, mail} = req.body;
+
+        if (!firstname || !lastname || !mail){
+            res.status(400).json({message: 'champs manquants'})
+            return;
+        }
+
+        try {
+            const student = await  this.studentService.createStudent({firstname, lastname, mail});
+            res.status(201).json(student);
+        }
+        catch (err){
+            console.error(err);
+            res.status(500).json({message:'Erreur de connexion à la base'});
+        }
+    };
+
+     async updateStudent (req:Request, res:Response){
         const id = Number(req.params.id);
         if (isNaN(id)){
             res.status(400).json({message:'id invalide'});
+            return;
+        }
+        try {
+            const student = await  this.studentService.updateStudent(id, req.body);
+            res.status(200).json(student);
+        }catch (err){
+            console.error(err);
+            res.status(404).json({ message: 'id pas trouvé' });
+        }
+
+    }
+
+    async deleteStudent (req:Request, res:Response){
+        const id = Number(req.params.id);
+        if (isNaN(id)){
+            res.status(400).json({message:'id invalide'});
+            return;
         }
         try {
             const studentsDeleted = await this.studentService.deleteStudent(id)
